+++ host/cpu/cpu_other.go
