+++ host/host_arm.go
@@ -10,6 +10,8 @@ import (
 	_ "github.com/google/periph/host/bcm283x"
 	_ "github.com/google/periph/host/chip"
 	_ "github.com/google/periph/host/odroid_c1"
+	// While this board is ARM64, it may run ARM 32 bits binaries so load it on
+	// 32 bits builds too.
 	_ "github.com/google/periph/host/pine64"
 	_ "github.com/google/periph/host/rpi"
 )
